code: check rate limit before each pull request
Previously, the rate limit was checked before a link was run. However,
if a link performed a bunch of operations, the link might still run out
of api requests (ie, if there are a lot of forks).

I extracted out all the rate checking logic in a function and am now
calling it before making each pull request. Whihc still not perfect,
it's a lot better then before.